Add a flag to suppress empty compressed chunks
When using zstr with boost::serialization, sync() is called at the
beginning, leading to an 8 byte empty compressed chunk. I want to
avoid this chunk as it breaks compatibility with the combination of
boost::compression and boost::iostream. Thus I introduce the
allow_empty_sync methods and flag (defaults to true) that may be used
to suppress these empty chunks.